Refactor MongoDB Compass download functionality into Tools package; update main.go to utilize new method
